centralize dynamic columns recovery and use typelist
