+++ Plugins/Podio/include/Acts/Plugins/Podio/PodioUtil.hpp
@@ -11,16 +11,21 @@
 #include "Acts/EventData/SourceLink.hpp"
 #include "Acts/Geometry/GeometryIdentifier.hpp"
 #include "Acts/Geometry/TrackingGeometry.hpp"
+#include "Acts/Plugins/Podio/PodioDynamicColumns.hpp"
+#include "Acts/Utilities/HashedString.hpp"
 #include "Acts/Utilities/Helpers.hpp"
 
 #include <limits>
 #include <memory>
 
+#include <podio/Frame.h>
+
 namespace ActsPodioEdm {
 class Surface;
 }
 
-namespace Acts::PodioUtil {
+namespace Acts {
+namespace PodioUtil {
 
 using Identifier = uint64_t;
 constexpr Identifier kNoIdentifier = std::numeric_limits<Identifier>::max();
@@ -42,5 +47,15 @@ std::shared_ptr<const Surface> convertSurfaceFromPodio(
 
 ActsPodioEdm::Surface convertSurfaceToPodio(const ConversionHelper& helper,
                                             const Acts::Surface& surface);
+}  // namespace PodioUtil
 
-}  // namespace Acts::PodioUtil
+namespace podio_detail {
+/// This is used by both the track and track state container, so the
+/// implementation is shared here
+void recoverDynamicColumns(
+    const podio::Frame& frame, const std::string& stem,
+    std::unordered_map<HashedString,
+                       std::unique_ptr<podio_detail::ConstDynamicColumnBase>>&
+        dynamic);
+}  // namespace podio_detail
+}  // namespace Acts
